Adjusted mosh_bloom
Now mosh_bloom waits a certain number of frames, blooms for a certain
number of frames, then skips the bloomed frames and goes to the
original point on the gif.  We could change this to instead go back to
the beginning fairly easily.
Future progress: Have our mosh programs put the gif through mosh.py
before messing with them, so we can get gifs with only i-frames fairly
easily.